+++ frontend/src/components/Main.jsx
@@ -21,6 +21,7 @@ import EditRestrauntModal from './modal/EditRestrauntModal';
 import { ShowRestrauntModal } from './modal/ShowRestrauntModal';
 import CreateReviewModal from './modal/CreateReviewModal';
 import {TagList} from './TagList';
+import {DateTimeConverter} from './DateTimeConverter'
 
 
 // import Restaurants from './../restaurants.json';
@@ -423,20 +424,6 @@ export const Main = (props) => {
     return nameFilter && isTagSelected
   })
 
-  const DateTimeConverter = (props) => {
-    const inputDate = props.created_at;
-    const date = new Date(inputDate);
-  
-    const year = date.getFullYear();
-    const month = ("0" + (date.getMonth() + 1)).slice(-2);
-    const day = ("0" + date.getDate()).slice(-2);
-    const hours = ("0" + date.getHours()).slice(-2);
-    const minutes = ("0" + date.getMinutes()).slice(-2);
-    const seconds = ("0" + date.getSeconds()).slice(-2);
-    const formattedDateTime = `${year}/${month}/${day} ${hours}:${minutes}:${seconds}`;
-    
-    return <div>{formattedDateTime}</div>;
-  };
 
 
   return (
